app/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow==10.3.0